osa4 valmis

+++ osa4/controllers/blogs.js
@@ -1,33 +1,62 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user')
     response.json(blogs.map(blog => blog.toJSON()))
   })
   
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
 
+    try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return response.status(401).json( {error: 'token missing or invalid '})
+        }
+    
+
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog({
         "title": body.title,
         "author": body.author,
         "url": body.url,
-        "likes": body.likes || 0
+        "likes": body.likes || 0,
+        "user": user.id
     })
   
-    try {
-        const savedBlog = await blog.save()
-        response.json(savedBlog.toJSON())
+    
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog.id)
+    await user.save()
+    response.json(savedBlog.toJSON())
+    
     } catch(exception) {
         next(exception)
     }
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
+    try{
+        
+        decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return response.status(401).json( {error: 'token missing or invalid '})
+        }
+
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(request.params.id)
+        
+        if ( user.id.toString() === blog.user.toString() ) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            response.status(401).json( {error: "wrong id"} )
+        }
+
     } catch (exception) {
         next(exception)
     }
